Increasing live (sync) operation timeout to 45 seconds.
The sync integration tests were using a timeout of 30 seconds for live
operations. Some of the sync stress tests are timing out because it
takes longer for a full sync cycle to complete.

This patch increases the timeout to 45 seconds in
profile_sync_service_harness.cc, and is expected to eliminate the
occasional failures on the builders.

TBR=tim@chromium.org
BUG=68207
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6125006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@71058 0039d316-1c4b-4281-b951-d872f2087c98