Add PrerenderResourceHandler and hook it into the ResourceDispatcherHost.
The PrerenderResourceHandler will initiate prerendering of a web page under the following conditions:
  - The initial request is a GET for a ResourceType::PREFETCH resource.
  - The mime-type (sniffed or explicitly specified) of the resource is text/html.
  - The response status code is a 200.
  - The top-level page will not need to be revalidated until after it's guaranteed to be removed from potential use.

The handler passes along all data to the backing resource handler.

BUG=61745
TEST=unit_tests --gtest_filter="PrerenderResourceHandlerTest*"

Manual Test: Start Chrome with --enable-page-prerender. Go to a page with a <link rel=prefetch> element. Navigate to that element, and make sure that it uses the prerendered view.

Review URL: http://codereview.chromium.org/5912001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@71162 0039d316-1c4b-4281-b951-d872f2087c98