Make thumbnails closable in tabpose
Also make the thumbnails left-aligned. The reason for this is that when thumbnails are closed, the reshuffling algorithm looks a bit busy. We want to eventually slide thumbs out on the left side on the window and then slide them back in from the right.

BUG=50307
TEST=
* Hover a thumb. It should show a close button.
* Press opt. All thumbs should show a close button.
* Press a close button. Thumb should close, the rest of the tabs should reshuffle.

Review URL: http://codereview.chromium.org/5960008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@71232 0039d316-1c4b-4281-b951-d872f2087c98