Simplified PrefValueStore::NotifyPrefChanged.
Reason for the refactoring is the need to properly handle disappearing policies.
This situation arises when a managed preference is removed that only has
been registered in one pref store (for example local_store) the notification
is sent to all pref stores though but the profile pref store doesn't have
a default value for that pref and therefore ControllingPrefStoreForPref returns
INVALID_STORE which shouldn't be treated as an exceptional situation (DCHECKed).

BUG=71028
TEST=unit_tests --gtest_filter=*Pref* should give pretty good coverage.

Review URL: http://codereview.chromium.org/6357019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@72787 0039d316-1c4b-4281-b951-d872f2087c98