Implement the DownloadDirectory policy.
This policy allows for locking the user to download files only to the specified location.

When the policy is set the UI elements should be inactive and the save dialog should never appear.

Changes in the Preferences.xib file only wire the enabled status of the download location and save prompt checkbox elements to a variable that tracks the managed state of the DefaultDownloadDir pref. No UI elements were harmed in this change.

BUG=59768
TEST=All old download manager tests shouldn't break. UI should behave correctly when policy is set.

Review URL: http://codereview.chromium.org/6525054

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@75255 0039d316-1c4b-4281-b951-d872f2087c98