Create dom_login watch list and added myself to it
This is a watch list for all of the files that are core to the DOM based login. There has been a couple of occassions where people have refactored this code not realizing that it is under very active development and thus causing issues with CLs that are currently in the air.

One of the rules, specifically 'chrome/browser/chromeos/login/.*dom.*' is meant for files that have not been created yet in the repo. The CL (6729018) for creating these files should be landing in the next day or two, so I didn't want to go through the bother of having multiple CLs for a change I know is coming.

Patch from Ryan Harrison <rharrison@chromium.org>

BUG=none
TEST=manual


Review URL: http://codereview.chromium.org/6765029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@80284 0039d316-1c4b-4281-b951-d872f2087c98