sync: Make nudge + config jobs reliable in SyncerThread2
We save nudges that are received during config. And execute them when mode changes to normal.

We also save nudge and config jobs that fail due to any reason. Then we execute them either during exponential backoff or when we get connection.

In short we are making the nudges and configs reliable.
BUG=77820
TEST=

Review URL: http://codereview.chromium.org/6812004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@81701 0039d316-1c4b-4281-b951-d872f2087c98