Increase timeout values for pyauto sync tests.
The pyauto sync test that restarts the browser was timing out on the
pyauto bots, due to the default timeout value of 20 seconds.

This patch increases the timeout for sync tests to 1 minute (for the
basic sign in test) and 2 minutes (for tests that go through more than
one sync cycle).

It also adds debugging information to the error message for cases where
sync does not complete within the wait period.

BUG=79894
TEST=sync.SyncTest.testRestartBrowser

Review URL: http://codereview.chromium.org/6883052

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@82147 0039d316-1c4b-4281-b951-d872f2087c98