sync: null check DTM in OnPassphraseAccepted
If this is called before OnBackendInitialized, the DTM won't exist.  We fake a call to OnPassphraseAccepted from OnPassphraseRequired if no encrypted types are enabled to clean up errors and passphrase prompts, but we don't need to call in to DTM.  Especially if it doesn't exist.

BUG=80208
TEST=Restart sync-enabled browser that doesn't have password sync enabled.  No crash.

Review URL: http://codereview.chromium.org/6880117

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@82550 0039d316-1c4b-4281-b951-d872f2087c98