Fix glitchy drawing in faded tab titles
When fading tab titles we would sometimes have some glitches when transitioning from the solid title to the faded title.

The problem was that we were switching the front from normal clear type to cleartype off. This changed caused visible glitches when transitioning between the two fonts.

Initially I switched clear type off to make it easier to calculate the alpha of the text. msw recomended using the luminosity of the text to determine the alpha. It turns out that this works great and we no longer have to turn off clear type in the faded sections.

BUG=78869
TEST=Reproduced the bug. Applied my patch and verified that the bug no longer reproduced.

Review URL: http://codereview.chromium.org/6902016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@82972 0039d316-1c4b-4281-b951-d872f2087c98