Coupling AllowFileSelectionDialogs-Pollicy and PromptForDownload.
When file-selection dialogs are forbidden by policy it is better to force the PromptForDownload preference to false and automatically download to the default location, since otherwise the user would just be prompted with a message saying that file-selection dialogs are disallowed.

Also cleaned up ConfigurationPolicyPrefStore* tests.

BUG=73174
TEST=Manually set the policy to false and verify that the PromptForDownload preference is false and cannot be turned on. Added ConfigurationPolicyPrefStorePromptDownloadTest.*.


Review URL: http://codereview.chromium.org/6927008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@84279 0039d316-1c4b-4281-b951-d872f2087c98